create edit delete word_answer

+++ application/controllers/Word.php
@@ -9,6 +9,7 @@ class Word extends My_Controller
         $this->lang->load('word', 'fels');
         $this->lang->load('home', 'fels');
         $this->lang->load('session', 'fels');
+        $this->lang->load('category', 'fels');
     }
 
     public function index($page = 1) 
@@ -32,11 +33,20 @@ class Word extends My_Controller
         $this->load->view('layout/index', $data);
     }
 
+    public function show($id) 
+    {
+        $data['word'] = $fag = $this->Word_Model->get_word(array('id' => $id));
+        $data['word_answer'] = $word_answer = $this->Word_Answer_Model->get_answer(array('word_id' => $fag['id'], 'correct' => 1));
+        $data['title'] = lang('word');
+        $data['template'] = 'word/show';
+        $data['authentication'] = $this->authentication;
+        $this->load->view('layout/index', $data);
+    }
+
     public function add() 
     {
         if ($this->input->post('add_word')) {
-            $this->form_validation->set_rules('content', lang('word'), 'trim|required' );
-            $this->form_validation->set_error_delimiters('<div style="color:red">','</div>');
+            $this->set_rules();
 
             if($this->form_validation->run()) {
                 $array = array(
@@ -44,8 +54,10 @@ class Word extends My_Controller
                     'content' => $this->input->post('content'),
                 );
                 $fag = $this->Word_Model->insert($array);
+                $checkbox = $this->input->post('checkbox');
+                $check = $this->set_word_answer(NULL, $checkbox, $fag['insert_id'], 0);
 
-                if ($fag > 0) {
+                if ($fag['affected_rows'] > 0) {
                     $fag = array(
                         'type' => 'successful',
                         'message' => lang('add_word_successful'),
@@ -60,13 +72,16 @@ class Word extends My_Controller
                 redirect('words');	
             }
         }
-        // cái này em lấy dữ liệu để test thôi. chứ sau thì phải lấy từ bảng category ra ạ
-        $data['list_category'] = array(
-            '1' => 'category1',
-            '2' => 'category2',
-            '3' => 'category3',
-            '4' => 'category4',
-        );
+        $total = $this->Category_Model->total();
+        $list_category = $this->Category_Model->view_category(0, $total);
+        $list[0] = lang('choose_catrgory');
+
+        if (isset($list_category)&& count($list_category)) {
+            foreach ($list_category as $key => $value) {
+                $list[$value['id']] = $value['name'];
+            }
+        }
+        $data['list_category'] = $list;
         $data['meta_title'] = lang('title_add_word');
         $data['template'] = 'word/add';
         $data['authentication'] = $this->authentication;
@@ -75,7 +90,8 @@ class Word extends My_Controller
 
     public function edit($id = 0) 
     {
-        $data['word'] = $word = $this->Word_Model->get($id);
+        $data['word'] = $this->Word_Model->get_word(array('id' => $id));
+        $data['list_word_answer'] = $list_word_answer = $this->Word_Answer_Model->get_answer(array('word_id' => $id));
 
         if (!isset($data['word']) || count($data['word']) == 0) {
             $fag = array(
@@ -87,23 +103,24 @@ class Word extends My_Controller
         }
 
         if ($this->input->post('edit_word')) {
-            $this->form_validation->set_rules('content', lang('word'), 'trim|required' );
-            $this->form_validation->set_error_delimiters('<div style="color:red">','</div>');
-
-            if($this->form_validation->run()) {
+            $this->set_rules();
+            
+            if ($this->form_validation->run()) {
                 $array = array(
                     'category_id' => $this->input->post('category'),
                     'content' => $this->input->post('content'),
                 );
                 $fag = $this->Word_Model->update($id, $array);
+                $checkbox = $this->input->post('checkbox');
+                $check = $this->set_word_answer($list_word_answer, $checkbox, $id, 1);
 
-                if ($fag > 0) {
-                    $fag = array(
+                if ($fag > 0 || $check > 0) {
+                    $fag = array (
                         'type' => 'successful',
                         'message' => lang('edit_word_successful'),
                     );
                 } else {
-                    $fag = array(
+                    $fag = array (
                         'type' => 'error',
                         'message' => lang('edit_word_error'),
                     );
@@ -112,12 +129,16 @@ class Word extends My_Controller
                 redirect('words');
             }
         }
-        $data['list_category'] = array(
-            '1' => 'category1',
-            '2' => 'category2',
-            '3' => 'category3',
-            '4' => 'category4',
-        );
+        $total = $this->Category_Model->total();
+        $list_category = $this->Category_Model->view_category(0, $total);
+        $list[0] = lang('choose_catrgory');
+
+        if (isset($list_category)&& count($list_category)) {
+            foreach ($list_category as $key => $value) {
+                $list[$value['id']] = $value['name'];
+            }
+        }
+        $data['list_category'] = $list;
         $data['meta_title'] = lang('title_edit_word');
         $data['template'] = 'word/edit';
         $data['authentication'] = $this->authentication;
@@ -126,9 +147,10 @@ class Word extends My_Controller
 
     public function delete($id = 0) 
     {
-        $data['word'] = $word = $this->Word_Model->get($id);
+        $word = $this->Word_Model->get_word(array('id' => $id));
+        $list_word_answer = $this->Word_Answer_Model->get_answer_id(array('word_id' => $word['id']));
 
-        if (!isset($data['word']) || count($data['word']) == 0) {
+        if (!isset($word) || count($word) == 0) {
             $fag = array(
                 'type' => 'error',
                 'message' => lang('no_word'),
@@ -137,6 +159,7 @@ class Word extends My_Controller
             redirect('words');
         }
         $fag = $this->Word_Model->delete((array)$id);
+        $fag = $this->Word_Answer_Model->delete($list_word_answer);
 
         if ($fag > 0) {
             $fag = array(
@@ -153,12 +176,20 @@ class Word extends My_Controller
         redirect('words');
     }
 
-    public function _action() 
+    protected function _action() 
     {
         if($this->input->post('delete_more')) {
             $checkbox = $this->input->post('checkbox');
-            $fag = $this->Word_Model->delete($checkbox);
-            
+
+            if (isset($checkbox) && count($checkbox)) {
+
+                foreach ($checkbox as $key => $value) {
+                    $list_word_answer = $this->Word_Answer_Model->get_answer_id(array('word_id' => $value));       
+                }
+                $fag = $this->Word_Model->delete($checkbox);
+                $fag = $this->Word_Answer_Model->delete($list_word_answer);
+            }
+                        
             if ($fag > 0) {
             $fag = array(
                 'type' => 'successful',
@@ -173,5 +204,41 @@ class Word extends My_Controller
             $this->session->set_flashdata('message_flashdata', $fag);
             redirect('words');        
         }        
+    }
+
+    protected function set_rules()
+    {
+        $this->form_validation->set_rules('content', lang('word'), 'trim|required' );
+        $this->form_validation->set_rules('answer[]', lang('answer'), 'trim|required' );
+        $this->form_validation->set_error_delimiters('<div style="color:red">','</div>');
+    }
+
+    protected function set_word_answer ($param_where, $list_checkbox, $id ,$fag)
+    {
+        $check = 0;
+        for ($i = 0; $i < 4; $i++) {
+            if (in_array($i, $list_checkbox)) {
+                $array_word_answer = array(
+                    'content' => $this->input->post('answer['.$i.']'),
+                    'word_id' => $id,  
+                    'correct' => 1,
+                );
+            } else {
+                $array_word_answer = array(
+                    'content' => $this->input->post('answer['.$i.']'),
+                    'word_id' => $id,
+                    'correct' => 0,                        
+                );
+            }
+            
+            if ($fag == 1) {
+                $fag1 = $this->Word_Answer_Model->update($param_where[$i]['id'], $array_word_answer);
+                if ($fag1 > 0) $check = 1;    
+            } else {
+                $fag1 = $this->Word_Answer_Model->insert($array_word_answer);
+                if ($fag1 > 0) $check = 1;
+            }
+        }
+        return $check;
     }
 }
